pyproject.toml
[tool.poetry]
name = "youtube_uploader"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
google-auth-oauthlib = "^0.8.0"
google-auth = "^2.15.0"
google-api-python-client = "^2.70.0"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
black = "^22.12.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
youtube_uploader = "youtube_uploader.main:main"